move p2p logic into seperate files
